Huge cleanup refactoring, remove androMda, move groovy scripts, remove unused classes and files, remove remnants of R connectivity, refactor DesigElement services and DAO, refactor outlier detection and cormat computation, etc.

+++ gemma-core/src/main/java/ubic/gemma/core/apps/DeleteExperimentsCli.java
@@ -1,13 +1,13 @@
 /*
  * The Gemma project
- * 
+ *
  * Copyright (c) 2011 University of British Columbia
- * 
+ *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
  * the License. You may obtain a copy of the License at
- * 
+ *
  * http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
@@ -19,43 +19,41 @@ import ubic.gemma.model.expression.experiment.ExpressionExperiment;
 
 /**
  * Delete one or more experiments from the system.
- * 
- * @author paul
  *
+ * @author paul
  */
 public class DeleteExperimentsCli extends ExpressionExperimentManipulatingCLI {
 
     public static void main( String[] args ) {
         DeleteExperimentsCli d = new DeleteExperimentsCli();
-
-        d.doWork( args );
+        Exception e = d.doWork( args );
+        if ( e != null ) {
+            e.printStackTrace();
+        }
     }
 
-    /*
-     * (non-Javadoc)
-     * 
-     * @see ubic.gemma.core.util.AbstractCLI#getCommandName()
-     */
     @Override
     public String getCommandName() {
         return "deleteExperiments";
     }
 
     @Override
-    public String getShortDesc() {
-        return "Delete experiment(s) from the system";
-    }
-
-    @Override
     protected Exception doWork( String[] args ) {
         this.force = true;
-        super.processCommandLine( args );
+        Exception e = super.processCommandLine( args );
+        if ( e != null )
+            return e;
 
         for ( BioAssaySet bas : this.expressionExperiments ) {
             this.eeService.remove( ( ExpressionExperiment ) bas );
         }
 
         return null;
+    }
+
+    @Override
+    public String getShortDesc() {
+        return "Delete experiment(s) from the system";
     }
 
 }
